rafactor - my-bundles type 폴더 분리 및 type명 구체화 - type 필드명 카멜케이스로 변경
