+++ src/app/my-bundles/page.tsx
@@ -18,10 +18,10 @@ import {
   DrawerTitle,
   DrawerClose,
 } from "@/components/ui/drawer";
+import { useHandleCreateBundleClick } from "@/hooks/bundle/add/useHandleCreateBundleClick";
 import { useDeleteMyBundleMutation } from "@/queries/useDeleteMyBundleMutation";
 import { useMyBundlesQuery } from "@/queries/useMyBundlesQuery";
-import { MyBundle } from "@/types/bundle/types";
-import { useHandleCreateBundleClick } from "@/hooks/bundle/add/useHandleCreateBundleClick";
+import { MyBundle } from "@/types/my-bundles/types";
 
 const Page = () => {
   const [isEdit, setIsEdit] = useState(false);
@@ -99,8 +99,8 @@ const Page = () => {
                 <MyBundleCard
                   key={bundle.id}
                   isEdit={isEdit}
-                  design_type={bundle.designType}
-                  is_read={bundle.isRead}
+                  designType={bundle.designType}
+                  isRead={bundle.isRead}
                   status={bundle.status}
                   name={bundle.name}
                   updatedAt={bundle.updatedAt}
@@ -117,8 +117,8 @@ const Page = () => {
                   <Link key={bundle.id} href={`/my-bundles/${bundle.id}`}>
                     <MyBundleCard
                       isEdit={isEdit}
-                      design_type={bundle.designType}
-                      is_read={bundle.isRead}
+                      designType={bundle.designType}
+                      isRead={bundle.isRead}
                       status={bundle.status}
                       name={bundle.name}
                       updatedAt={bundle.updatedAt}
